Add cast of value to size_t.
git-svn-id: svn://svn.code.sf.net/p/bogofilter/code/trunk/bogofilter@1997 046ba1d7-8659-4f4f-8137-08d775d845e8

+++ src/common.h
@@ -47,7 +47,7 @@
 #define max(x, y)	(((x) > (y)) ? (x) : (y))
 #define min(x, y)	(((x) < (y)) ? (x) : (y))
 
-#define COUNTOF(array)	(sizeof(array)/sizeof(array[0]))
+#define COUNTOF(array)	(size_t)(sizeof(array)/sizeof(array[0]))
 
 typedef unsigned char byte;
